Upstream release v0.9.8 (#38)
* Update P2P demos

* Only send (invite) events to the AS if it's interested in it (#2692)

This should hopefully fix #2691 (and maybe also #2690)

* Move `SetLatestEvents` call

* Optimise `resolve-state` tool

* State and auth tweaks (update to matrix-org/gomatrixserverlib#335)

* Tweak logging for detected state resets

* Improve Sentry reporting

* Set Sentry user where known

* Tweak Sentry again

* State resolution fixes in `applyAndAuthEvent` (update to matrix-org/gomatrixserverlib#336)

* Stronger default power levels

* Update to matrix-org/gomatrixserverlib@637a173

* Handle errors differently in the `DeviceListUpdater` (#2695)

`If a device list update goes missing, the server resyncs on the next
one` was failing because a previous test would receive a `waitTime` of
1h, resulting in the test timing out.
This now tries to handle the returned errors differently, e.g. by using
the default `waitTime` of 2s. Also doesn't try further users in the
list, if one of the errors would cause a longer `waitTime`.

* Add a SigningKeyUpdate producer (#2697)

This adds a new stream for signing key updates, this should ensure we
don't lose any updates over federation.

* Re-add waitTime if we're not blacklisted and no RetryAfter was
specified.

* Avoid unneeded JSON operations (#2698)

We were `json.Unmarshal`ing the EDU and `json.Marshal`ing right before
sending the EDU to the stream. Those are now removed and the consumer
does `json.Unmarshal` once.

* Add HTTP status code to FederationClientError (#2699)

Also ensures we wait on more HTTP status codes.

* Update state reset capture to Sentry

* Use a single context instead

* Fulltext implementation incl. config (#2480)

This adds the main component of the fulltext search.
This PR doesn't do anything yet, besides creating an empty fulltextindex
folder if enabled. Indexing events is done in a separate PR.

* Fix issue with stale device lists (#2702)

We were only sending the last entry to the worker, so most likely missed
updates.

* Update getting pushrules, add tests, tweak pushrules (#2705)

This PR
- adds tests for `evaluatePushrules`
- removes the need for the UserAPI on the `OutputStreamEventConsumer`
(for easier testing)
- adds a method to get the pushrules from the database
- adds a new default pushrule for `m.reaction` events (and some other
tweaks)

* Use database writer in user API `DeleteOldNotifications`

* More writer fixes in the Sync API

* Change detection of already executed migrations (#2665)

This changes the detection of already executed migrations for the
roomserver state block and keychange refactor. It now uses schema tables
provided by the database engine to check if the column was already
removed. We now also store the migration in the migrations table.

This should stop e.g. Postgres from logging errors like `ERROR: column
"event_nid" does not exist at character 8`.

* Fix transactions around assigning NIDs

* Fix database transaction for keyserver `DeleteDeviceKeys`

* Optimise shared users query in Synx API slightly by removing a potential sort

* Optimise `SharedUsers` again by using complete composite index

* Version 0.9.7 (#2707)

Changelog and version bump.

* Only verify if fulltext indexing is enabled

* Update to matrix-org/gomatrixserverlib@dec87dbaa40700a8c8f7c299c53f911047cf2cb9

* P2P demo tweaks

* Update to matrix-org/pinecone@b215925

* Update to NATS Server v2.9.0

* Fix missing signature check on the `/get_missing_events` response

* Version 0.9.8

Co-authored-by: Neil Alexander <neilalexander@users.noreply.github.com>
Co-authored-by: Till <2353100+S7evinK@users.noreply.github.com>